Refactor legacy_authenticate.php
Previusly password_verify pass throught a if statement to determinate if
the exit was 0 (without errors) or 1 (with errors). The refactor version
pass the negative boolean result to a variable and there is changed the data
type to int and pass to the exit method.
This changed is thanks to "Your Common Sense" from stack exchange